requirements.txt
h5py==3.6.0
nltk==3.6.7
opencv-python==4.5.5.62
pandas==1.3.5
scikit-learn==1.0.2
tensorboard==2.7.0
torch==1.8.1
torchtext==0.5.0
torchvision==0.9.1
matplotlib==3.5.1
jupyter==1.0.0
ipywidgets==7.6.5
stanza==1.3.0
PyYAML==6.0
